move inference to its own place
