+++ libs/wingc/src/type_check.rs
@@ -1,4 +1,5 @@
 mod class_fields_init;
+mod inference_visitor;
 pub(crate) mod jsii_importer;
 pub mod lifts;
 pub mod symbol_env;
@@ -9,7 +10,6 @@ use crate::ast::{
 	Interface as AstInterface, InterpolatedStringPart, Literal, Phase, Reference, Scope, Spanned, Stmt, StmtKind, Symbol,
 	TypeAnnotation, UnaryOperator, UserDefinedType,
 };
-use crate::closure_transform::CLOSURE_CLASS_PREFIX;
 use crate::comp_ctx::{CompilationContext, CompilationPhase};
 use crate::diagnostic::{report_diagnostic, Diagnostic, TypeError, WingSpan};
 use crate::docs::Docs;
@@ -34,6 +34,7 @@ use wingii::fqn::FQN;
 use wingii::type_system::TypeSystem;
 
 use self::class_fields_init::VisitClassInit;
+use self::inference_visitor::InferenceVisitor;
 use self::jsii_importer::JsiiImportSpec;
 use self::lifts::Lifts;
 use self::symbol_env::{LookupResult, LookupResultMut, SymbolEnvIter, SymbolEnvRef};
@@ -221,135 +222,6 @@ pub enum Type {
 	Interface(Interface),
 	Struct(Struct),
 	Enum(Enum),
-}
-
-struct InferenceVisitor<'a> {
-	types: &'a mut Types,
-	expected_type: Option<&'a TypeRef>,
-	found_inference: bool,
-}
-
-impl<'a> crate::visit_types::VisitTypeMut<'_> for InferenceVisitor<'a> {
-	fn visit_typeref_mut(&mut self, node: &'_ mut TypeRef) {
-		if matches!(**node, Type::Inferred(_)) {
-			*node = self.types.maybe_unwrap_inference(*node);
-		} else {
-			crate::visit_types::visit_typeref_mut(self, node);
-		}
-	}
-
-	// structs and interfaces cannot have inferences
-	fn visit_interface_mut(&mut self, _node: &'_ mut Interface) {}
-	fn visit_struct_mut(&mut self, _node: &'_ mut Struct) {}
-
-	fn visit_class_mut(&mut self, node: &'_ mut Class) {
-		// We only care about visiting a class if it represent an inflight closure, where inference is possible.
-		// In which case, we need to visit the function signature of the handle method
-		if node.name.name.starts_with(CLOSURE_CLASS_PREFIX) {
-			if let Some(handle) = node.env.lookup_mut(&Symbol::global(CLOSURE_CLASS_HANDLE_METHOD), None) {
-				self.visit_typeref_mut(&mut handle.as_variable_mut().unwrap().type_);
-			}
-		}
-	}
-
-	fn visit_function_signature_mut(&mut self, node: &'_ mut FunctionSignature) {
-		for param in node.parameters.iter_mut() {
-			self.visit_typeref_mut(&mut param.typeref);
-		}
-
-		self.visit_typeref_mut(&mut node.return_type);
-	}
-}
-
-impl<'a> crate::visit_types::VisitType<'_> for InferenceVisitor<'a> {
-	fn visit_typeref(&mut self, node: &'_ TypeRef) {
-		if let Some(expected) = self.expected_type {
-			match &**expected {
-				// unwrap the simple "wrapper" types to correspond to diving into the actual type we're looking at
-				Type::Optional(t)
-				| Type::Array(t)
-				| Type::MutArray(t)
-				| Type::Map(t)
-				| Type::MutMap(t)
-				| Type::Set(t)
-				| Type::MutSet(t) =>
-				// If the type we're looking at is also a wrapper type, then we need to unwrap it
-				{
-					match &**node {
-						Type::Optional(_)
-						| Type::Array(_)
-						| Type::MutArray(_)
-						| Type::Map(_)
-						| Type::MutMap(_)
-						| Type::Set(_)
-						| Type::MutSet(_) => {
-							self.expected_type = Some(t);
-						}
-						_ => {}
-					}
-				}
-
-				Type::Anything
-				| Type::Number
-				| Type::String
-				| Type::Duration
-				| Type::Boolean
-				| Type::Void
-				| Type::Json
-				| Type::MutJson
-				| Type::Nil
-				| Type::Unresolved
-				| Type::Function(_)
-				| Type::Class(_)
-				| Type::Interface(_)
-				| Type::Struct(_)
-				| Type::Enum(_) => {}
-
-				Type::Inferred(_) => {
-					// Inferences are not a useful expected type
-					self.expected_type = None;
-				}
-			}
-		}
-
-		crate::visit_types::visit_typeref(self, node);
-	}
-
-	// structs and interfaces cannot have inferences
-	fn visit_interface(&mut self, _node: &'_ Interface) {}
-	fn visit_struct(&mut self, _node: &'_ Struct) {}
-
-	fn visit_class(&mut self, node: &'_ Class) {
-		// We only care about visiting a class if it represent an inflight closure, where inference is possible.
-		// In which case, we need to visit the function signature of the handle method
-		if let Some(method) = node.get_closure_method() {
-			self.visit_typeref(&method);
-		}
-	}
-
-	fn visit_function_signature(&mut self, node: &'_ FunctionSignature) {
-		let expected_function_sig = if let Some(ref expected) = self.expected_type {
-			expected.as_deep_function_sig()
-		} else {
-			None
-		};
-
-		for (idx, param) in node.parameters.iter().enumerate() {
-			self.expected_type = expected_function_sig.and_then(|f| f.parameters.get(idx).map(|p| &p.typeref));
-			self.visit_typeref(&param.typeref);
-		}
-
-		self.expected_type = expected_function_sig.map(|f| &f.return_type);
-		self.visit_typeref(&node.return_type);
-	}
-
-	fn visit_inference(&mut self, node: &'_ usize) {
-		self.found_inference = true;
-
-		if let Some(expected) = self.expected_type {
-			self.types.update_inferred_type(*node, *expected);
-		}
-	}
 }
 
 pub const CLASS_INIT_NAME: &'static str = "init";
